rafactor: reflect review changes
